examples/simple-reporting/requirements.txt
t2g-sdk==1.0.0rc7
simple_graph_retriever==1.0.0-rc.5
google-generativeai
